requirements.txt
absl-py==0.15.0
anyio==3.6.2
asgiref==3.4.1
astunparse==1.6.3
attrs==19.3.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==2.0.12
clang==5.0
class-doc==0.2.0b0
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
fastapi==0.83.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h11==0.13.0
h5py==3.1.0
idna==3.10
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
more-itertools==7.2.0
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.9.2
pydantic-settings==0.2.5
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2026.1.post1
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.1.0
tomlkit==0.5.11
tqdm==4.64.1
typing-extensions==4.1.1
urllib3==1.26.20
uvicorn==0.16.0
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
